requirements.txt
nosetests
feedparser
